django-supermarket-peoject/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
django==4.0.3
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
python-decouple==3.6
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
